dialog paths instead of queries
